[SIMPLY-2855] Fix a crash when selecting a library as a new user.
